Seperate goods from agent for better documentation
