+++ docs/Goods.rst
@@ -10,7 +10,7 @@ An agent can access a good with :code:`self['cookies']` or
 - :code:`self.create(money, 15)` creates money
 - :code:`self.destroy(money, 10)` destroys money
 - goods can be given, taken, sold and bought
-- :code:`self['money'] returns the quantity an agent possesses
+- :code:`self['money']` returns the quantity an agent possesses
 
 Services
 --------
@@ -19,7 +19,14 @@ Services are like goods, but the need to be declared as services
 in the simulation :func:`abce.__init__.service`.
 In this function one declares a good that creates the other good and
 how much. For example if one has :code:`self['adults'] = 2`, one could
-get 16 hours of labor every day. :code:`simulation.declare_service('adults', 8, 'labor').
+get 16 hours of labor every day. :code:`simulation.declare_service('adults', 8, 'labor')`.
 
 
+.. default-domain::abce.Goods
 
+.. automodule:: abce.goods
+
+.. autoclass:: abce.goods.Goods
+    :members:
+    :undoc-members:
+    :show-inheritance:
